Split diagnostic builders into src/lcm/regime_building/diagnostics.py.
Move `_build_compute_intermediates_per_period`, `_wrap_with_reduction`,
and `_productmap_over_state_action_space` out of `processing.py` into
a dedicated `diagnostics.py` module.

The diagnostic builder is a cold-path artifact: it only runs when
`validate_V` detects NaN. Its sibling builders in `processing.py`
(`_build_Q_and_F_per_period`, `_build_max_Q_over_a_per_period`) are
on the hot path, invoked every backward-induction step. Grouping by
frequency-of-concern is more useful here than grouping by
time-of-execution. `processing.py` shrinks by ~200 lines.

`get_compute_intermediates` stays in `Q_and_F.py` — it is a structural
mirror of `get_Q_and_F` and sharing privates (`_get_U_and_F`,
`_get_arg_names_of_Q_and_F`, `_get_joint_weights_function`). History
shows they can drift (ec96c60 "Sync compute_intermediates with
get_Q_and_F"); colocation makes the next drift visible in the same
file.

Runtime-side error handling (`_enrich_with_diagnostics`,
`_summarize_diagnostics`, `_format_diagnostic_summary`) stays in
`lcm.utils.error_handling` — that's where the exception enrichment
legitimately lives.
